backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
httpx>=0.28.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
bcrypt>=4.2.0